тестик

+++ src/ru/job4j/condition/SqArea.java
@@ -1,7 +1,7 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(int p, int k) {
+    public static double square(int p, double k) {
    double h = p / (2 * (k + 1));
    return Math.pow(h, 2) * k;
